Move label_column_for_table column lookup into SqliteIntrospector
Closes seam 5. The SQLite-specific part of label_column_for_table -- reading a table's columns and single-column unique indexes via sqlite_utils -- becomes Introspector.column_details_with_uniqueness(). The label-picking heuristic and config lookup (table_config) stay on Database, which now consults the introspector. sqlite_utils import dropped from database.py.

Behaviour-preserving; tests pass unchanged.
